Added spanish translation Updated code so it compiles with gcc 3.4 Added some implementation for effect framework, but cannot remember exactly how far I had got previously
svn path=/trunk/kdenlive/; revision=368

+++ kdenlive/kmmrulerpanel.cpp
@@ -99,6 +99,6 @@ void KMMRulerPanel::sliderScaleChange(int value)
 
 void KMMRulerPanel::setScale(double scale)
 {
-	int newScale = maxFrameSize / scale;
+	double newScale = maxFrameSize / scale;
 	m_scaleSlider->setValue((int)(log(newScale / expA) / expK));
 }
